RTL Support first submit - WIP
This is a work in progress of Rocket.Chat RTL Support. The work is not
finished. I didn't know where to put the .rtl class in the html file, so
I use the addClass method in layout.coffee. Please write down suggestion
and don't accept this pull request yet, until final version.